pyproject.toml
[tool.poetry]
name = "studatio"
version = "0.1.9"
description = "Personal tool for my violin teaching database"
authors = ["Eliza Wilson <elizaaverywilson@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/elizaaverywilson/studatio"


[tool.poetry.dependencies]
python = ">=3.9, <3.11"
pyperclip = "^1.8.2"
tomlkit = "^0.11.4"
appdirs = "^1.4.4"

[tool.poetry.group.vendorizedicalevents.dependencies]
icalendar = "4.0.9"
python-dateutil = "^2.8.2"
pytz = "^2021.3"
DateTime = "^4.3"
httplib2 = "^0.20.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.3"
coverage = "^6.5.0"
vendorize = "^0.3.0"
pytest-datadir = "^1.3.1"

[tool.poetry.scripts]
studatio = 'studatio.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
